Fix stride type in process_range.

+++ source/mango/jpeg/jpeg_decode.cpp
@@ -2365,9 +2365,9 @@ namespace jpeg {
     void Parser::finishProgressiveST()
     {
         const size_t stride = m_surface->stride;
-        const int bytes_per_pixel = m_surface->format.bytes();
-        const int xstride = bytes_per_pixel * xblock;
-        const int ystride = stride * yblock;
+        const size_t bytes_per_pixel = m_surface->format.bytes();
+        const size_t xstride = bytes_per_pixel * xblock;
+        const size_t ystride = stride * yblock;
 
         u8* image = m_surface->address<u8>(0, 0);
 
@@ -2442,9 +2442,9 @@ namespace jpeg {
         const int yblock_last = yclip ? yclip : yblock;
 
         const size_t stride = m_surface->stride;
-        const int bytes_per_pixel = m_surface->format.bytes();
-        const int xstride = bytes_per_pixel * xblock;
-        const int ystride = stride * yblock;
+        const size_t bytes_per_pixel = m_surface->format.bytes();
+        const size_t xstride = bytes_per_pixel * xblock;
+        const size_t ystride = stride * yblock;
 
         u8* image = m_surface->address<u8>(0, 0);
